useGlobalState: pr changes
1. overloaded useDeepCompareRef
2. moved enhanceDispatch outside of hook
3.  uncluttered useGlobalState file by moving code out
4. applyMiddlware inside middleware folder
5. cleanup
